refactor: [FE] 결과스케치북에 하단 컴포넌트 리팩터링
