+++ frontend/src/components/resultSketchbook/ResultSketchbook.tsx
@@ -3,48 +3,32 @@ import toast, { Toaster } from 'react-hot-toast';
 import styled from 'styled-components';
 import { Center } from '@styles/styled';
 import { useRecoilValue } from 'recoil';
-import {
-    canOneMoreGameState,
-    currentBookIdxState,
-    gameResultIdState,
-    isStartedState,
-    maxSketchbookState,
-    sketchbookAuthorState,
-} from '@atoms/result';
-import { userState } from '@atoms/user';
+import { canOneMoreGameState, gameResultIdState, isStartedState } from '@atoms/result';
 import { lobbyIdState } from '@atoms/game';
-import { ReactComponent as LeftArrowIcon } from '@assets/icons/chevron-left-gradient.svg';
-import { ReactComponent as RightArrowIcon } from '@assets/icons/chevron-right-gradient.svg';
 import { ReactComponent as ExportIcon } from '@assets/icons/export-icon.svg';
 import resultInSound from '@assets/sounds/result-in.wav';
 import { networkServiceInstance as NetworkService } from '@services/socketService';
-import { emitOneMoreGame } from '@game/NetworkServiceUtils';
 import useMovePage from '@hooks/useMovePage';
 import useCheckGuidePage from '@hooks/useCheckGuidePage';
-import useResultSketchbook from '@hooks/useResultSketchbook';
 import useCopyClipBoard from '@hooks/useCopyClipboard';
 import useSoundEffect from '@hooks/useSoundEffect';
-import PrimaryButton from '@components/PrimaryButton';
 import SketchbookCard from '@components/SketchbookCard';
 import ResultGuide from '@components/resultSketchbook/ResultGuide';
 import QuizResultContent from '@components/resultSketchbook/QuizResultContent';
 import QuizAuthor from '@components/resultSketchbook/QuizAuthor';
 import CurSketchbookPage from '@components/resultSketchbook/CurSketchbookPage';
+import SketchbookAuthor from '@components/resultSketchbook/SketchbookAuthor';
+import OneMoreGameButton from '@components/resultSketchbook/OneMoreGameButton';
 
 function ResultSketchbook({ isForShareResult }: { isForShareResult: boolean }) {
     const [setPage] = useMovePage();
     const lobbyId = useRecoilValue(lobbyIdState);
     const gameResultId = useRecoilValue(gameResultIdState);
-    const { maxBookNum } = useRecoilValue(maxSketchbookState);
-    const sketchbookAuthor = useRecoilValue(sketchbookAuthorState);
-    const currentBookIdx = useRecoilValue(currentBookIdxState);
 
-    const { isHost } = useRecoilValue(userState);
     const isStarted = useRecoilValue(isStartedState);
     const canOneMoreGame = useRecoilValue(canOneMoreGameState);
 
     const { checkIsNotGuidePage } = useCheckGuidePage();
-    const { changeSketchbook } = useResultSketchbook(isForShareResult);
     const [_, onCopy] = useCopyClipBoard();
     const { playSoundEffect } = useSoundEffect();
 
@@ -84,32 +68,10 @@ function ResultSketchbook({ isForShareResult }: { isForShareResult: boolean }) {
                     )
                 }
             />
-            <SketchbookAuthor>
+            <OutsideOfCard>
                 {!isStarted && (
                     <>
-                        {isHost && currentBookIdx !== 0 && (
-                            <LeftArrowIcon
-                                onClick={() => changeSketchbook(-1)}
-                                role={'button'}
-                                aria-label={'이전 유저 스케치북 보기'}
-                            />
-                        )}
-                        {isHost && currentBookIdx === 0 && <EmptySpan />}
-                        <Brace>{'{'}</Brace>
-                        <SketchbookAuthorName isHost={isHost}>
-                            {sketchbookAuthor}
-                        </SketchbookAuthorName>
-                        <Brace>{'}'}</Brace>
-                        <span>의 스케치북</span>
-                        {isHost && currentBookIdx === maxBookNum && <EmptySpan />}
-                        {isHost && currentBookIdx !== maxBookNum && (
-                            <RightArrowIcon
-                                onClick={() => changeSketchbook(1)}
-                                role={'button'}
-                                aria-label={'다음 유저 스케치북 보기'}
-                            />
-                        )}
-
+                        <SketchbookAuthor isForShareResult={isForShareResult} />
                         <ButtonWrapper>
                             {(isForShareResult || canOneMoreGame) && (
                                 <ExportIcon
@@ -118,26 +80,18 @@ function ResultSketchbook({ isForShareResult }: { isForShareResult: boolean }) {
                                     aria-label={'게임 결과 페이지 링크 복사'}
                                 />
                             )}
-                            {!isForShareResult && canOneMoreGame && isHost && (
-                                <div
-                                    onClick={emitOneMoreGame}
-                                    role={'button'}
-                                    aria-label={'게임 한판 더 하기'}
-                                >
-                                    <PrimaryButton topText='ONE MORE' bottomText='한판 더 하기' />
-                                </div>
-                            )}
+                            <OneMoreGameButton isForShareResult={isForShareResult} />
                         </ButtonWrapper>
                     </>
                 )}
-            </SketchbookAuthor>
+            </OutsideOfCard>
         </>
     );
 }
 
 export default ResultSketchbook;
 
-const SketchbookAuthor = styled(Center)`
+const OutsideOfCard = styled(Center)`
     width: 100%;
     height: 65px;
     position: relative;
@@ -154,26 +108,6 @@ const SketchbookAuthor = styled(Center)`
     > span {
         font-weight: 600;
     }
-`;
-
-const SketchbookAuthorName = styled.span<{ isHost: boolean | null }>`
-    background: ${(props) =>
-        props.isHost ? props.theme.gradation.yellowPurple : props.theme.color.whiteT2};
-    ${({ theme }) => theme.layout.gradientTypo}
-`;
-
-const Brace = styled.span`
-    margin: 0 6px;
-    font-family: 'Sniglet', cursive;
-    font-weight: 700;
-    font-size: ${({ theme }) => theme.typo.h1};
-    transform: translateY(4px);
-`;
-
-const EmptySpan = styled.span`
-    width: 30px;
-    margin: 0 28px;
-    transform: scale(1.3) translateY(2px);
 `;
 
 const ButtonWrapper = styled(Center)`
